Add files via upload
The STABILO Digipen Reference Data Tool is a Python 3 script. The program has a main window which is vertically split in two halves. All user entry takes place in the right half. A protocol window informs the user of the most recent results. The sequence of operating the program requires three simple steps. The respective buttons are activated only when the previous step has been completed.

The program consists of these source files and their dependencies:
 - main.py Should be called when run from the terminal. Needed imports: sys
 - classes.py Defines some common structures. Needed imports: dataclass from dataclasses, pandas, date from datetime
 - ui.py Controls the user interface. Needed imports: os.path, shutil, PySimpleGUI, platform, traceback, pandas
 - file.py Handles all input and output. Needed imports: os, csv, traceback, numpy, pandas, add from operator, date from datetime, pickle
 - data.py Does the main computational tasks. Needed imports: numpy, pandas
 - plot.py Opens a separate window for x-y-plots of selected results. Needed imports: numpy, platform, matplotlib, savgol_filter from scipy.signal

The purpose of the program is to compile writer recordings into the parameters used as reference in the EduPen app. All recordings should be collected in one main directory. Writer recordings should be collected in this directory or subdirectories within it.